requirements.txt
amqp==2.2.1
appdirs==1.4.3
Babel==2.4.0
beautifulsoup4==4.5.3
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
click==6.7
Flask==0.12.2
Flask-Admin==1.5.0
flask-allows==0.2.0
Flask-BabelEx==0.9.3
Flask-BabelPlus==2.1.0
Flask-Caching==1.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
flask-mongoengine==0.9.2
Flask-Plugins==1.6.1
Flask-Redis==0.3.0
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
mongoengine==0.11.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pymongo==3.4.0
pyparsing==2.2.0
python-memcached==1.58
pytz==2017.2
qiniu==7.1.4
redis==2.10.5
requests==2.13.0
six==1.10.0
speaklater==1.3
vine==1.1.4
Werkzeug==0.12.2
WTForms==2.1
